fix(integrity): ZIP組立と内容検証を分離
Refs: #2092
